Add files via upload
This is a updated version to fix the bug. Because if you use airmon-ng it puts your interface as interfacemon
I was missing the mon and fixed it. There is a few thing in it that make no scence but I was sending one of my friends the program so he could work on a local exploit. For wpa2