Add files via upload
Added a new crunch command. I plan to add a keyspace program that you just enter NETGEAR and uses the keyspace for that router. But it will come soon when I get bored. I also added better tabs more variables and touched up the program. I need to add the # and comments but I dont care because well I know what it does.